Moved the reporting_db instance to be global to the PipelineBase
We basically just had too many instances of connecting to it individually, so we're just going to make it part of the common pipeline instance. Even though we don't use it in every pipeline, we're just gunna make it common for now, unless there arises a good justification to make it an option later.